% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_data.R
\name{create_sample_data}
\alias{create_sample_data}
\title{Title}
\usage{
create_sample_data(
  treatment = c("TRT1", "TRT2"),
  light_isotope = c("12C"),
  heavy_isotope = c("13C"),
  replicate = c("A", "B", "C"),
  timepoint = c("early")
)
}
\arguments{
\item{timepoint}{}
}
\description{
Title
}
